refactor constants and hook to buttons/common.ts
